fix(value): move tests to tests.rs
