+++ boa/src/builtins/value/tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::{forward, Executor, Realm};
 
 #[test]
 fn check_is_object() {
@@ -45,4 +46,51 @@ fn check_number_is_true() {
     assert_eq!(Value::from(-0.0).is_true(), false);
     assert_eq!(Value::from(-1.0).is_true(), true);
     assert_eq!(Value::from(NAN).is_true(), false);
+}
+
+// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Equality_comparisons_and_sameness
+#[test]
+fn abstract_equality_comparison() {
+    let realm = Realm::create();
+    let mut engine = Executor::new(realm);
+
+    assert_eq!(forward(&mut engine, "undefined == undefined"), "true");
+    assert_eq!(forward(&mut engine, "null == null"), "true");
+    assert_eq!(forward(&mut engine, "true == true"), "true");
+    assert_eq!(forward(&mut engine, "false == false"), "true");
+    assert_eq!(forward(&mut engine, "'foo' == 'foo'"), "true");
+    assert_eq!(forward(&mut engine, "0 == 0"), "true");
+    assert_eq!(forward(&mut engine, "+0 == -0"), "true");
+    assert_eq!(forward(&mut engine, "+0 == 0"), "true");
+    assert_eq!(forward(&mut engine, "-0 == 0"), "true");
+    assert_eq!(forward(&mut engine, "0 == false"), "true");
+    assert_eq!(forward(&mut engine, "'' == false"), "true");
+    assert_eq!(forward(&mut engine, "'' == 0"), "true");
+    assert_eq!(forward(&mut engine, "'17' == 17"), "true");
+    assert_eq!(forward(&mut engine, "[1,2] == '1,2'"), "true");
+    assert_eq!(forward(&mut engine, "new String('foo') == 'foo'"), "true");
+    assert_eq!(forward(&mut engine, "null == undefined"), "true");
+    assert_eq!(forward(&mut engine, "undefined == null"), "true");
+    assert_eq!(forward(&mut engine, "null == false"), "false");
+    assert_eq!(forward(&mut engine, "[] == ![]"), "true");
+    assert_eq!(
+        forward(&mut engine, "a = { foo: 'bar' }; b = { foo: 'bar'}; a == b"),
+        "false"
+    );
+    assert_eq!(
+        forward(&mut engine, "new String('foo') == new String('foo')"),
+        "false"
+    );
+    assert_eq!(forward(&mut engine, "0 == null"), "false");
+
+    // https://github.com/jasonwilliams/boa/issues/357
+    assert_eq!(forward(&mut engine, "0 == '-0'"), "true");
+    assert_eq!(forward(&mut engine, "0 == '+0'"), "true");
+    assert_eq!(forward(&mut engine, "'+0' == 0"), "true");
+    assert_eq!(forward(&mut engine, "'-0' == 0"), "true");
+
+    // https://github.com/jasonwilliams/boa/issues/393
+    // assert_eq!(forward(&mut engine, "0 == NaN"), "false");
+    // assert_eq!(forward(&mut engine, "'foo' == NaN"), "false");
+    // assert_eq!(forward(&mut engine, "NaN == NaN"), "false");
 }
